Refactor: Extract expression-position bindings to reactive-data.js
Step 5 per ai/plans/native-renderer-blocks.md. Biggest line-count win in
the decomposition — collapses four methods and their hydration duplicates
into one module.

- reactive-data.js: four plain functions grouped by dispatch on
  entry.classification.type:
    • bindAttribute        — property / event / boolean / ifDefined /
                             single-expression / interpolated-string
    • bindTextExpression   — unsafeHTML / literalValue / default text
    • bindRawTextContent   — script/style/textarea/title content
    • hydrateTextExpression — adopts server DOM with splitText boundary
  These are NOT blocks — no DynamicRegion, no owned region. Inline
  bindings with one Reaction per marker.
- renderer.js: four Renderer methods collapse to one-line delegations;
  skipFirstWrite is passed through for hydration. Functions imported as
  *Fn to avoid name collision with the delegating method wrappers

1003 tests pass (173 unit + 751 renderer browser + 79 component).
Attribute-bindings (42) and subtree-spurious (22) confirm reactive
attribute behavior, ssr-hydration (73) confirms skipFirstWrite + split
text adoption on hydrate.
